[shared] move CrossHair back to xy-chart and fix linting + tests
