+++ packages/xy-chart/src/chart/XYChart.jsx
@@ -3,9 +3,10 @@ import PropTypes from 'prop-types';
 
 import Grid from '@vx/grid/build/grids/Grid';
 import Group from '@vx/group/build/Group';
+import WithTooltip, { withTooltipPropTypes } from '@data-ui/shared/build/enhancer/WithTooltip';
 
 import Voronoi from './Voronoi';
-import WithTooltip, { withTooltipPropTypes } from '@data-ui/shared/build/enhancer/WithTooltip';
+
 
 import {
   collectDataFromChildSeries,
